+++ ImpactfulSkills/common/Util.cs
@@ -1,6 +1,7 @@
 ﻿using HarmonyLib;
 using System.Collections.Generic;
 using System.Reflection.Emit;
+using UnityEngine;
 
 namespace ImpactfulSkills.common
 {
@@ -20,6 +21,21 @@ namespace ImpactfulSkills.common
             matcher.Labels.Clear();
 
             return matcher;
+        }
+
+        public static List<ZNetPeer> ServerGetPeersInArea(Vector3 pos, float radius) {
+            var result = new List<ZNetPeer>();
+            if (!ZNet.instance || !ZNet.instance.IsServer())
+                return result;
+
+            float radiusSqr = radius * radius;
+            foreach (ZNetPeer peer in ZNet.instance.m_peers) {
+                if (!peer.IsReady() || peer.m_characterID == ZDOID.None)
+                    continue;
+                if (Utils.DistanceSqr(peer.m_refPos, pos) <= radiusSqr)
+                    result.Add(peer);
+            }
+            return result;
         }
     }
 }